backend/requirements.txt
Django==6.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
django-cors-headers==4.3.1
django-filter==24.1
mysqlclient==2.2.8
cryptography==42.0.5
redis==5.0.3
celery==5.3.6
django-celery-beat==2.9.0
Pillow>=10.2.0
gunicorn==21.2.0
python-dotenv==1.0.1
requests==2.31.0
